requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.4
django-ckeditor==6.0.0
django-crispy-forms==1.11.2
django-js-asset==1.2.2
gunicorn==20.0.4
Pillow==8.0.1
psycopg2==2.8.6
python-dotenv==0.18.0
python-http-client==3.3.1
pytz==2020.4
sendgrid==6.4.8
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
whitenoise==5.2.0
